initial conversion to class structure.
